higher level implementation of healthz checker (#7)
## Details 

The following patch updates the healthz handler feature from expecting a handler to expecting a collection of Service Checkers.
A Service Checker is a function which request the status of a service that the current service rely on. The status can take three values: `healthy`, `degraded`, `unhealthy`, the status of the current service will take the worst value returned by its checkers.