implement the types in go pkg (#44)
Made `database.DB` type an interface and ported the former implementation so we can use it almost as is in our code. 
The test is a bit messy and I don't really like it, but I couldn't come up with a cleaner implementation, if you have ideas, suggestions are welcomed.